+++ src/ImageCaching.Nuke/ApiDefinition.cs
