+++ src/ImageCaching.Nuke/Structs.cs
