+++ src/Sample/AppDelegate.cs
@@ -1,7 +1,7 @@
 ﻿using Foundation;
 using UIKit;
 
-namespace Xamarin.Nuke.Sample
+namespace Sample
 {
     // The UIApplicationDelegate for the application. This class is responsible for launching the
     // User Interface of the application, as well as listening (and optionally responding) to application events from iOS.
