+++ src/Sample/Assets.xcassets/AppIcon.appiconset/Contents.json
