+++ src/Sample/Assets.xcassets/AppIcon.appiconset/Icon1024.png
