+++ src/Sample/Assets.xcassets/AppIcon.appiconset/Icon120.png
