+++ src/Sample/Assets.xcassets/AppIcon.appiconset/Icon152.png
