+++ src/Sample/Assets.xcassets/AppIcon.appiconset/Icon40.png
