+++ src/Sample/Assets.xcassets/AppIcon.appiconset/Icon58.png
