+++ src/Sample/Assets.xcassets/AppIcon.appiconset/Icon60.png
