+++ src/Sample/Assets.xcassets/AppIcon.appiconset/Icon76.png
