+++ src/Sample/Assets.xcassets/Placeholder.imageset/Contents.json
