+++ src/Sample/Assets.xcassets/Placeholder.imageset/baseline_account_circle_black_60pt_3x.png
