+++ src/Sample/Resources/LaunchScreen.xib
